+++ lib/chronic/handlers.rb
@@ -317,6 +317,15 @@ module Chronic
       end
     end
 
+    def handle_sm_rmn_sy(tokens, options)
+      day = tokens[0].get_tag(ScalarDay).type
+      month = tokens[1].get_tag(RepeaterMonthName).index
+      year = tokens[2].get_tag(ScalarYear).type
+      time = Chronic.time_class.local(year, month, day)
+      end_time = Chronic.time_class.local(year, month, day + 1)
+      Span.new(time, end_time)
+    end
+
     # anchors
 
     # Handle repeaters