+++ HISTORY.md
@@ -3,6 +3,7 @@
 * Start using minitest for testing
 * Ensure periods are interpreted as colons (#81).
 * Support month/day and day/month parsing (#59).
+* Support day(scalar)-month(name)-year(scalar) (#99).
 
 # 0.6.7 / 2012-01-31